rename: pages에 세그먼트 추가
