+++ src/pages/chat/ui/layout.vue
